refactor: change app component naming & move schdule into home
