+++ apps/web/src/app/(causw)/alumni-contacts/page.tsx
@@ -1,5 +1,5 @@
 import { AlumniContactsPage } from '@/_pages/alumni-contacts';
 
-export default function Home() {
+export default function Page() {
   return <AlumniContactsPage />;
 }
