+++ apps/web/src/app/(causw)/home/page.tsx
@@ -1,5 +1,5 @@
 import { HomePage } from '@/_pages/home';
 
-export default function Home() {
+export default function Page() {
   return <HomePage />;
 }
