+++ apps/web/src/app/(causw)/setting/page.tsx
@@ -1,5 +1,5 @@
 import { SettingPage } from '@/_pages/setting';
 
-export default function Home() {
+export default function Page() {
   return <SettingPage />;
 }
