+++ apps/web/src/widgets/home/model/index.ts
@@ -1,0 +1,1 @@
+export * from './schedule';
