+++ apps/web/src/widgets/home/model/schedule/constants.ts
