+++ apps/web/src/widgets/home/model/schedule/index.ts
