+++ apps/web/src/widgets/home/model/schedule/types.ts
