+++ apps/web/src/widgets/home/ui/index.ts
@@ -5,3 +5,4 @@ export * from './HomeHeader';
 export * from './HomeNotificationCard';
 export * from './HomeQuickMenu';
 export * from './HomeScheduleList';
+export * from './schedule';
