+++ apps/web/src/widgets/home/ui/schedule/index.ts
@@ -1,0 +1,1 @@
+export * from './ScheduleListWidget';
