--- apps/web/src/widgets/schedule/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './SchduleListWidget';
